Refactor Okta Verify flow to share the same base view. OKTA-390965
